Linux: Drop fragment headers in favour of a total length header
Now that we no longer have to deal with fragments of different messages
interleaving, there is no need for the complicated fragment ID handling.
The only information we strictly need is whether we have fragmentation
at all. (So the receiver knows whether to retrieve and use the dedicated
channel.)

However, rather than just sending a boolean flag, we can just as well
send the total size of the message. This way the receiver knows when
fragmentation occurs -- and additionally knows exactly when we have
received enough packets. This avoids the need to check in the receiver
for the channel being closed by the sender; and knowing the total size
in advance will also enable further optimisations/simplifications in the
future.

On the sender side, avoiding the fragment headers removes the need to
copy the data for each fragment -- this doubles performance (send +
receive) of large (fragmented) transfers.

On the receiver side, a similar optimisation should be possible;
however, unlike for the sender, this will require a major reorganisation
of the code -- this calls for a separate optimisation commit.