Linux: Drop fragment headers in favour of a total length header
Now that we no longer have to deal with fragments of different messages
interleaving, there is no need for the complicated fragment ID handling.
The only information we strictly need is whether we have fragmentation
at all. (So the receiver knows whether to retrieve and use the dedicated
channel.)

However, rather than just sending a boolean flag, we can just as well
send the total size of the message. This way the receiver knows when
fragmentation occurs -- and additionally knows exactly when we have
received enough packets. This avoids the need to check in the receiver
for the channel being closed by the sender; and knowing the total size
in advance will also enable further optimisations/simplifications in the
future.

On the sender side, avoiding the fragment headers removes the need to
copy the data for each fragment -- this almost doubles performance (send
+ receive) of large transfers.

On the receiver side, a similar optimisation should be possible;
however, unlike for the sender, this will require a major reorganisation
of the code -- this calls for a separate optimisation commit.

   test _invoke_chaos ... bench:   3,173,412 ns/iter (+/- 621,316)
   test size_00_1     ... bench:      11,616 ns/iter (+/- 108)
   test size_01_2     ... bench:      11,628 ns/iter (+/- 96)
   test size_02_4     ... bench:      11,342 ns/iter (+/- 100)
   test size_03_8     ... bench:      11,376 ns/iter (+/- 102)
   test size_04_16    ... bench:      11,412 ns/iter (+/- 80)
   test size_05_32    ... bench:      11,471 ns/iter (+/- 88)
   test size_06_64    ... bench:      11,668 ns/iter (+/- 83)
   test size_07_128   ... bench:      11,977 ns/iter (+/- 110)
   test size_08_256   ... bench:      12,469 ns/iter (+/- 134)
   test size_09_512   ... bench:      13,314 ns/iter (+/- 135)
   test size_10_1k    ... bench:      15,053 ns/iter (+/- 251)
   test size_11_2k    ... bench:      18,340 ns/iter (+/- 187)
   test size_12_4k    ... bench:      25,112 ns/iter (+/- 819)
   test size_13_8k    ... bench:      41,653 ns/iter (+/- 1,578)
   test size_14_16k   ... bench:      73,528 ns/iter (+/- 2,987)
   test size_15_32k   ... bench:     137,692 ns/iter (+/- 5,857)
   test size_16_64k   ... bench:     265,955 ns/iter (+/- 11,823)
   test size_17_128k  ... bench:     509,746 ns/iter (+/- 22,675)
   test size_18_256k  ... bench:   1,110,666 ns/iter (+/- 409,253)
   test size_19_512k  ... bench:   1,754,337 ns/iter (+/- 563,647)
   test size_20_1m    ... bench:   3,966,698 ns/iter (+/- 519,277)
   test size_21_2m    ... bench:   7,872,532 ns/iter (+/- 406,799)
   test size_22_4m    ... bench:  12,792,789 ns/iter (+/- 682,725)
   test size_23_8m    ... bench:  24,867,029 ns/iter (+/- 1,628,664)