Linux: Use separate send buffers for header and main data
Using the scatter-gather functionality of the sendmsg() system call, we can
avoid the need for copying around data altogether. This boosts performance
(send + receive) of large transfers by another >50% or so, resulting in a total
speedup of about 2x - 4x (depending on size) compared to the original version.
Medium (non-fragemented) transfers also gain a few percent.

(The gains on the sender side are actually even bigger: we still haven't
optimised the receiver side at all -- so that is now the main bottleneck...)

Note that this change actually obsoletes the performance gains of the previous
change: using scatter-gather, we could have achieved the same zero-copy effect
even with the old fragment header approach, with only a very small overhead for
creating the headers. Still, the code is considerably simpler using the new
approach.

   test _invoke_chaos ... bench:   2,819,098 ns/iter (+/- 519,666)
   test size_00_1     ... bench:      11,665 ns/iter (+/- 397)
   test size_01_2     ... bench:      11,608 ns/iter (+/- 41)
   test size_02_4     ... bench:      11,535 ns/iter (+/- 64)
   test size_03_8     ... bench:      11,623 ns/iter (+/- 120)
   test size_04_16    ... bench:      11,615 ns/iter (+/- 99)
   test size_05_32    ... bench:      11,661 ns/iter (+/- 72)
   test size_06_64    ... bench:      11,751 ns/iter (+/- 101)
   test size_07_128   ... bench:      12,039 ns/iter (+/- 57)
   test size_08_256   ... bench:      12,469 ns/iter (+/- 107)
   test size_09_512   ... bench:      13,384 ns/iter (+/- 95)
   test size_10_1k    ... bench:      15,348 ns/iter (+/- 101)
   test size_11_2k    ... bench:      20,014 ns/iter (+/- 99)
   test size_12_4k    ... bench:      28,468 ns/iter (+/- 328)
   test size_13_8k    ... bench:      45,882 ns/iter (+/- 892)
   test size_14_16k   ... bench:      72,960 ns/iter (+/- 414)
   test size_15_32k   ... bench:     133,763 ns/iter (+/- 873)
   test size_16_64k   ... bench:     254,916 ns/iter (+/- 2,796)
   test size_17_128k  ... bench:     496,277 ns/iter (+/- 9,491)
   test size_18_256k  ... bench:     923,052 ns/iter (+/- 243,845)
   test size_19_512k  ... bench:   1,474,469 ns/iter (+/- 587,669)
   test size_20_1m    ... bench:   2,962,315 ns/iter (+/- 585,452)
   test size_21_2m    ... bench:   5,356,993 ns/iter (+/- 482,646)
   test size_22_4m    ... bench:   7,879,423 ns/iter (+/- 1,086,070)
   test size_23_8m    ... bench:  12,579,441 ns/iter (+/- 1,377,765)