Linux: Use separate send buffers for header and main data
Using the scatter-gather functionality of the sendmsg() system call, we
can avoid the need for copying around data altogether. This boosts
performance (send + receive) of large (fragmented) transfers by another
>50%, resulting in a total speedup of more than 3x compared to the
original version. Medium (non-fragemented) transfers gain >15% too.

(The gains on the sender side are actually much bigger: we still haven't
optimised the receiver side at all -- so that is now the main
bottleneck...)

Note that this change actually obsoletes the performance gains of the
previous change: using scatter-gather, we could have achieved the same
zero-copy effect even with the old fragment header approach, with only a
very small overhead for creating the headers. Still, the code is
considerably simpler using the new approach.