Cleanup: Linux: send: Move iovec creation out of construct_header()
While iovec is part of the msghdr, and it appears nicer in principle to
encapsulate its construction in the same helper function, that didn't
actually work, as we had to mutate the iovec later on anyways; so in the
end this only caused the same iovec struct unnecessarily being created
twice.

Moving iovec creation outside also avoids the need for the ugly Box<>
return hack -- so all in all it's probably cleaner to do it this way.