Linux: Use separate receive buffers for header and main message
Just like on the sender side, we can use the scatter-gather
functionality of recvmsg() to put the data directly into the final
place rather than having to copy it around, even for the initial
fragment.

This removes the last major piece of unnecessary overhead; and
consequently results in very large gains mostly for medium-sized
transfers, where the speedup exceeds 10x on my system. Larger transfers
of a few MiB are also affected quite significantly, with about 20 %
speedup.

All in all, transfers of all sizes are now several times faster than
with the unoptimised version -- with the lowest speedup on my system
at 4x for small transfers; about 4.5x for very large ones; and
the largest boost of >11x for medium-sized ones.

A quick check on a more modern system (64 bit; fairly recent Intel CPU)
showed even larger gains: while very big transfers were similar (about
5x speedup), small ones gained >9x, and medium-sized ones >20x.

One interesting observation is that on the modern system, for the
optimised version there are even more strongly pronounced jumps at 512
KiB and at 4 MiB. While I haven't veryfied whether these are more like
spikes or more like steps, I suspect it's the latter: with more
streamlined memory access patterns in the optimised version, it becomes
pretty obvious that these are simply successive levels of the cache
hierarchy being exhausted...

Below are the numbers of a typical run on the older system, as usual.

(Note that the numbers for 256 KiB and above are now significantly
affected by the benchmark having to launch an extra thread: without it,
the 256 KiB one weighs in at only 103,000 ns... For larger sizes, the
difference gets amortised of course. On the newer system on the other
hand launching the extra thread doesn't seem to have such a strongly
pronounced effect.)

   test size_00_1    ... bench:       2,742 ns/iter (+/- 45)
   test size_01_2    ... bench:       2,814 ns/iter (+/- 49)
   test size_02_4    ... bench:       2,858 ns/iter (+/- 37)
   test size_03_8    ... bench:       2,798 ns/iter (+/- 49)
   test size_04_16   ... bench:       2,817 ns/iter (+/- 62)
   test size_05_32   ... bench:       2,806 ns/iter (+/- 46)
   test size_06_64   ... bench:       2,792 ns/iter (+/- 77)
   test size_07_128  ... bench:       2,827 ns/iter (+/- 70)
   test size_08_256  ... bench:       2,951 ns/iter (+/- 67)
   test size_09_512  ... bench:       2,966 ns/iter (+/- 48)
   test size_10_1k   ... bench:       3,109 ns/iter (+/- 72)
   test size_11_2k   ... bench:       4,490 ns/iter (+/- 77)
   test size_12_4k   ... bench:       6,198 ns/iter (+/- 67)
   test size_13_8k   ... bench:      10,040 ns/iter (+/- 100)
   test size_14_16k  ... bench:       9,688 ns/iter (+/- 94)
   test size_15_32k  ... bench:      14,334 ns/iter (+/- 104)
   test size_16_64k  ... bench:      23,064 ns/iter (+/- 1,190)
   test size_17_128k ... bench:      45,455 ns/iter (+/- 2,803)
   test size_18_256k ... bench:     190,946 ns/iter (+/- 27,766)
   test size_19_512k ... bench:     379,753 ns/iter (+/- 55,044)
   test size_20_1m   ... bench:   1,200,600 ns/iter (+/- 140,428)
   test size_21_2m   ... bench:   1,555,805 ns/iter (+/- 275,438)
   test size_22_4m   ... bench:   4,703,751 ns/iter (+/- 653,501)
   test size_23_8m   ... bench:   9,370,280 ns/iter (+/- 2,133,685)