Linux: Don't check maximum send size individually for each channel
The buffer size shouldn't change from one channel to another; so instead
of using a syscall to check the size each time, just check it once using
a `lazy_static`.

Note that we only check the send size now, and use it for the size of
the receive buffer too. This is actually more correct than the previous
implementation, as the receive buffer needs to hold exactly as much as
can be sent. (Normally they are the same anyways; but if for some reason
the maximum receive size happened to be larger, the previous code would
use a larger buffer than necessary. If the receive size happened to be
*smaller*, either version would fail horribly...)

This doesn't have a noticable performance impact on the sender, as the
present implementation only checks the size *after* failing to send the
whole message in one packet, i.e. only for large transfers, where the
cost of the extra system call is insignificant. The receiver side on the
other hand always does the check; and thus the saved call actually
yields a significant improvement for small messages: on my system, small
transfers (send + receive) gain some 13%. Together with the other
improvement, they are now about 4.5x faster than the original
implementation.

   test size_00_1    ... bench:       2,379 ns/iter (+/- 73)
   test size_01_2    ... bench:       2,394 ns/iter (+/- 42)
   test size_02_4    ... bench:       2,493 ns/iter (+/- 36)
   test size_03_8    ... bench:       2,424 ns/iter (+/- 54)
   test size_04_16   ... bench:       2,456 ns/iter (+/- 52)
   test size_05_32   ... bench:       2,431 ns/iter (+/- 46)
   test size_06_64   ... bench:       2,396 ns/iter (+/- 48)
   test size_07_128  ... bench:       2,416 ns/iter (+/- 44)
   test size_08_256  ... bench:       2,498 ns/iter (+/- 37)
   test size_09_512  ... bench:       2,829 ns/iter (+/- 44)
   test size_10_1k   ... bench:       2,721 ns/iter (+/- 36)
   test size_11_2k   ... bench:       4,099 ns/iter (+/- 58)
   test size_12_4k   ... bench:       5,798 ns/iter (+/- 54)
   test size_13_8k   ... bench:       9,681 ns/iter (+/- 93)
   test size_14_16k  ... bench:       9,208 ns/iter (+/- 61)
   test size_15_32k  ... bench:      13,919 ns/iter (+/- 126)
   test size_16_64k  ... bench:      22,553 ns/iter (+/- 209)
   test size_17_128k ... bench:      45,079 ns/iter (+/- 3,714)
   test size_18_256k ... bench:     193,373 ns/iter (+/- 46,669)
   test size_19_512k ... bench:     371,403 ns/iter (+/- 48,139)
   test size_20_1m   ... bench:   1,354,696 ns/iter (+/- 190,947)
   test size_21_2m   ... bench:   1,524,060 ns/iter (+/- 219,842)
   test size_22_4m   ... bench:   4,746,800 ns/iter (+/- 512,016)
   test size_23_8m   ... bench:   9,389,120 ns/iter (+/- 772,502)