Linux: Refactor: Replace `construct_header()` by `send_first_fragment()`
Moving the actual sending of the message into the `construct_header()`
function, and renaming it to `send_first_fragment()` to reflect that
change.

With the previous changes, we always send the message right after
constructing the header -- so it makes sense to put the common sequence
in one place.

What's more, this helps isolating the unsafe operations: invoking
`send_first_fragment()` is in fact not unsafe at all.