inprocess: Use a `Drop` trait to free server records
Rather than freeing the records from `ONE_SHOT_SERVERS` when they are
"used up" on the receiver side (on `accept()`), clean them out when the
corresponding server handle (`OsIpcOneShotServer`) is dropped.

This seems more natural, and should avoid potential resource leaks (when
`accept()` is never called for some reason); but OTOH it requires
dealing with locks in a `drop()` implementation -- so I'm not entirely
sure this is actually a good idea...

Also, we need unit tests for the various possible situations here.