tests: cleanup: Avoid need for separate `_server()` pseudo-tests
Integrate the server functionality right into the main `_spawn()` test
functions: running either the server or the client portions as
necessary.

This further aligns the structure of these tests with the corresponding
`_fork()` variants. It also avoids the need for the
`#[ignore]`/`--ignored` hack, since all the test functions now always
run -- either as normal tests (client parts); or as pseudo-tests, when
self-spawned in server mode.