windows: Add checks and warnings regarding dangers of `read_buf`
The whole thing is highly unsafe. To handle it *properly*, we would need
to encapsulate it such that nothing else can access the buffer while
there is an outstanding async read in progress with the kernel...
However, that's quite an invasive change. So for now, the best we can do
is put adding big fat warnings everywhere, and additional safety checks
where possible.